+++ Changelog.md
@@ -25,6 +25,7 @@
 - Introduces uncategorized annotations grouping in Annotations settings tab (#4733)
 - Introduces new grades summary chart, and makes student view of grades consistent with admin (#4740)
 - Set SameSite=Lax on cookies (#4742)
+- Introduces individual marks chart view for assessments (#4747)
 - Fix annotation modal overflow issue (#4748)
 
 ## [v1.9.3]